added options for review

+++ FirstJobApp/src/main/java/com/emabrkx/FirstJobApp/company/impl/CompanyServiceImpl.java
@@ -33,10 +33,10 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public boolean deleteCompanyById(Long id) {
-        try {
+        if (companyRepository.existsById(id)) {
             companyRepository.deleteById(id);
             return true;
-        } catch (Exception e) {
+        } else {
             return false;
         }
     }
